requirements.txt
aiohttp
websockets
aiomysql
python-dateutil
pympler
flask
Telethon==1.11.3
redis
aioredis
hiredis
pillow
tgs
CairoSVG
aiofiles
python-ffmpeg
mutagen
hachoir
sanic